Struct opentelemetry_jaeger::PipelineBuilder

pub struct PipelineBuilder { /* private fields */ }

Jaeger exporter builder

Implementations

impl PipelineBuilder

pub fn with_agent_endpoint<T: ToSocketAddrs>(self, agent_endpoint: T) -> Self

Assign the agent endpoint.

pub fn with_instrumentation_library_tags(self, export: bool) -> Self

Config whether to export information of instrumentation library.

pub fn with_service_name<T: Into<String>>(self, service_name: T) -> Self

Assign the process service name.

pub fn with_tags<T: IntoIterator<Item = KeyValue>>(self, tags: T) -> Self

👎Deprecated since 0.16.0: please pass those tags as resource in sdk::trace::Config. Then use with_trace_config method to pass the config. All key value pairs in resources will be reported as process tags

Assign the process tags.

Note that resource in trace Config is also reported as process tags in jaeger. If there is duplicate tags between resource and tags. Resource’s value take priority even if it’s empty.

pub fn with_max_packet_size(self, max_packet_size: usize) -> Self

Assign the max packet size in bytes. Jaeger defaults is 65000.

pub fn with_auto_split_batch(self, auto_split: bool) -> Self

Config whether to auto split batches.

When auto split is set to true, the exporter will try to split the batch into smaller ones so that there will be minimal data loss. It will impact the performance.

Note that if one span is too large to export, other spans within the same batch may or may not be exported. In this case, exporter will return errors as we cannot split spans.

pub fn with_trace_config(self, config: Config) -> Self

Assign the SDK config for the exporter pipeline.

Examples

Set service name via resource.

use opentelemetry_jaeger::PipelineBuilder;
use opentelemetry::sdk;
use opentelemetry::sdk::Resource;
use opentelemetry::KeyValue;

let pipeline = PipelineBuilder::default()
                .with_trace_config(
                      sdk::trace::Config::default()
                        .with_resource(Resource::new(vec![KeyValue::new("service.name", "my-service")]))
                );

pub fn install_simple(self) -> Result<Tracer, TraceError>

Install a Jaeger pipeline with a simple span processor.

pub fn install_batch<R: JaegerTraceRuntime>( self, runtime: R ) -> Result<Tracer, TraceError>

Install a Jaeger pipeline with a batch span processor using the specified runtime.

pub fn build_simple(self) -> Result<TracerProvider, TraceError>

Build a configured sdk::trace::TracerProvider with a simple span processor.

pub fn build_batch<R: JaegerTraceRuntime>( self, runtime: R ) -> Result<TracerProvider, TraceError>

Build a configured sdk::trace::TracerProvider with a batch span processor using the specified runtime.

pub fn init_sync_exporter(self) -> Result<Exporter, TraceError>

Initialize a new simple exporter.

This is useful if you are manually constructing a pipeline.

pub fn init_async_exporter<R: JaegerTraceRuntime>( self, runtime: R ) -> Result<Exporter, TraceError>

Initialize a new exporter.

This is useful if you are manually constructing a pipeline.