Crate opentelemetry_jaeger

Collects OpenTelemetry spans and reports them to a given Jaeger agent or collector endpoint. See the Jaeger Docs for details about Jaeger and deployment information.

Compiler support: requires rustc 1.46+

Quickstart

First make sure you have a running version of the Jaeger instance you want to send data to:

$ docker run -d -p6831:6831/udp -p6832:6832/udp -p16686:16686 -p14268:14268 jaegertracing/all-in-one:latest

Then install a new jaeger pipeline with the recommended defaults to start exporting telemetry:

use opentelemetry::trace::Tracer;
use opentelemetry::global;

fn main() -> Result<(), opentelemetry::trace::TraceError> {
    global::set_text_map_propagator(opentelemetry_jaeger::Propagator::new());
    let tracer = opentelemetry_jaeger::new_pipeline().install_simple()?;

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });

    global::shutdown_tracer_provider(); // export remaining spans

    Ok(())
}

Performance

For optimal performance, a batch exporter is recommended as the simple exporter will export each span synchronously on drop. You can enable the [rt-tokio], [rt-tokio-current-thread] or [rt-async-std] features and specify a runtime on the pipeline builder to have a batch exporter configured for you automatically.

[dependencies]
opentelemetry = { version = "*", features = ["rt-tokio"] }
opentelemetry-jaeger = { version = "*", features = ["rt-tokio"] }

let tracer = opentelemetry_jaeger::new_pipeline()
    .install_batch(opentelemetry::runtime::Tokio)?;

Jaeger Exporter From Environment Variables

The jaeger pipeline builder can be configured dynamically via environment variables. All variables are optional, a full list of accepted options can be found in the jaeger variables spec.

Jaeger Collector Example

If you want to skip the agent and submit spans directly to a Jaeger collector, you can enable the optional collector_client feature for this crate. This example expects a Jaeger collector running on http://localhost:14268.

[dependencies]
opentelemetry-jaeger = { version = "..", features = ["collector_client", "isahc"] }

Then you can use the with_collector_endpoint method to specify the endpoint:

// Note that this requires the `collector_client` feature.
// We enabled the `isahc` feature for a default isahc http client.
// You can also provide your own implementation via new_pipeline().with_http_client() method.
use opentelemetry::trace::{Tracer, TraceError};

fn main() -> Result<(), TraceError> {
    let tracer = opentelemetry_jaeger::new_pipeline()
        .with_collector_endpoint("http://localhost:14268/api/traces")
        // optionally set username and password as well.
        .with_collector_username("username")
        .with_collector_password("s3cr3t")
        .install_simple()?;

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });

    Ok(())
}

Kitchen Sink Full Configuration

Example showing how to override all configuration options. See the PipelineBuilder docs for details of each option.

use opentelemetry::{KeyValue, trace::{Tracer, TraceError}};
use opentelemetry::sdk::{trace::{self, IdGenerator, Sampler}, Resource};
use opentelemetry::global;

fn main() -> Result<(), TraceError> {
    global::set_text_map_propagator(opentelemetry_jaeger::Propagator::new());
    let tracer = opentelemetry_jaeger::new_pipeline()
        .with_agent_endpoint("localhost:6831")
        .with_service_name("my_app")
        .with_tags(vec![KeyValue::new("process_key", "process_value")])
        .with_max_packet_size(9_216)
        .with_trace_config(
            trace::config()
                .with_sampler(Sampler::AlwaysOn)
                .with_id_generator(IdGenerator::default())
                .with_max_events_per_span(64)
                .with_max_attributes_per_span(16)
                .with_max_events_per_span(16)
                .with_resource(Resource::new(vec![KeyValue::new("key", "value")])),
        )
        .install_batch(opentelemetry::runtime::Tokio)?;

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });

    global::shutdown_tracer_provider(); // export remaining spans

    Ok(())
}

Crate Feature Flags

The following crate feature flags are available:

collector_client: Export span data directly to a Jaeger collector. User MUST provide the http client.

surf_collector_client: Export span data with Jaeger collector backed by a surf default http client.

reqwest_collector_client: Export span data with Jaeger collector backed by a reqwest http client.

reqwest_blocking_collector_client: Export span data with Jaeger collector backed by a reqwest blocking http client.

isahc_collector_client: Export span data with Jaeger collector backed by a isahc http client.

wasm_collector_client: Enable collector in wasm.

Support for recording and exporting telemetry asynchronously can be added via the following flags, it extends the opentelemetry feature:

rt-tokio: Enable sending UDP packets to Jaeger agent asynchronously when Tokio runtime is used.

rt-tokio-current-thread: Enable sending UDP packets to Jaeger agent asynchronously when [TokioCurrentThread] runtime is used.

rt-async-std: Enable sending UDP packets to Jaeger agent asynchronously when [AsyncStd] runtime is used.

Supported Rust Versions

OpenTelemetry is built against the latest stable release. The minimum supported version is 1.46. The current OpenTelemetry version is not guaranteed to build on Rust versions earlier than the minimum supported version.

The current stable Rust compiler and the three most recent minor versions before it will always be supported. For example, if the current stable compiler version is 1.49, the minimum supported version will not be increased past 1.46, three minor versions prior. Increasing the minimum supported compiler version is not considered a semver breaking change as long as doing so complies with this policy.

Structs

Exporter

Jaeger span exporter

PipelineBuilder

Jaeger exporter builder

Process

Jaeger process configuration

Propagator

The Jaeger propagator propagates span contexts in jaeger’s propagation format.

Enums

Error

Wrap type for errors from opentelemetry jaeger

Traits

JaegerTraceRuntime

Jaeger Trace Runtime is an extension to TraceRuntime. Currently it provides a UDP socket used by AgentAsyncClientUdp.

Functions

new_pipeline

Create a new Jaeger exporter pipeline builder.